pyproject.toml
[tool.poetry]
name = "swit-imap-satellite"
version = "0.1.0"
description = ""
authors = ["johnnykim <johnny.kim@swit.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
fastapi-sqlalchemy = "^0.2.1"
python-dotenv = "^1.0.1"
pymysql2 = "^1.3.3"
imap = "^1.0.0"
jinja2 = "^3.1.3"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pycryptodome = "^3.20.0"
gunicorn = "^22.0.0"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
